requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.9
environ==1.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
oauth2client==4.1.3
oauthlib==3.2.2
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-decouple==3.8
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
